+++ src/nl/hannahsten/texifyidea/run/linuxpdfviewer/OpenViewerListener.kt
@@ -8,7 +8,6 @@ import com.intellij.openapi.util.Key
 import nl.hannahsten.texifyidea.TeXception
 import nl.hannahsten.texifyidea.action.ForwardSearchAction
 import nl.hannahsten.texifyidea.run.latex.LatexRunConfiguration
-import nl.hannahsten.texifyidea.run.pdfviewer.ExternalPdfViewer
 import nl.hannahsten.texifyidea.run.pdfviewer.PdfViewer
 import org.jetbrains.concurrency.runAsync
 
@@ -21,7 +20,7 @@ class OpenViewerListener(
     private val sourceFilePath: String,
     val line: Int,
     val project: Project,
-    val focusAllowed: Boolean = true
+    val focusAllowed: Boolean = false
 ) :
     ProcessListener {
 
@@ -29,11 +28,15 @@ class OpenViewerListener(
         if (event.exitCode == 0) {
             runAsync {
                 try {
-                    when (viewer) {
-                        is InternalPdfViewer -> viewer.conversation?.forwardSearch(pdfPath = runConfig.outputFilePath, sourceFilePath = sourceFilePath, line = line, project = project, focusAllowed = focusAllowed)
-                        is ExternalPdfViewer -> viewer.forwardSearch(pdfPath = runConfig.outputFilePath, sourceFilePath = sourceFilePath, line = line, project = project, focusAllowed = focusAllowed)
-                        else -> {}
-                    }
+                    // ensure the viewer is open, especially for Sumatra
+                    viewer.openFile(runConfig.outputFilePath, project)
+                    viewer.forwardSearch(
+                        pdfPath = runConfig.outputFilePath,
+                        sourceFilePath = sourceFilePath,
+                        line = line,
+                        project = project,
+                        focusAllowed = focusAllowed
+                    )
                     // Set this viewer as viewer to forward search to in the future.
                     (ActionManager.getInstance().getAction("texify.ForwardSearch") as? ForwardSearchAction)?.viewer = viewer
                 }
@@ -41,9 +44,6 @@ class OpenViewerListener(
                 }
             }
         }
-
-        // Reset to default
-        runConfig.allowFocusChange = true
     }
 
     override fun onTextAvailable(event: ProcessEvent, outputType: Key<*>) {